move game options methods to /game options
